Refactor health in core component to its own component. Make health a usize where we can since that is simpler than float arithmetic. 

Spawn text effects to show damage numbers and picked up consumables.

+++ src/collision/mod.rs
@@ -33,7 +33,7 @@ pub enum SortedCollisionEvent {
         player_entity: Entity,
         projectile_entity: Entity,
         projectile_faction: Faction,
-        projectile_damage: f32,
+        projectile_damage: usize,
     },
     PlayerToConsumableIntersection {
         player_entity: Entity,
@@ -43,22 +43,22 @@ pub enum SortedCollisionEvent {
         player_entity: Entity,
         mob_entity: Entity,
         mob_faction: Faction,
-        player_damage: f32,
-        mob_damage: f32,
+        player_damage: usize,
+        mob_damage: usize,
     },
     PlayerToMobSegmentContact {
         player_entity: Entity,
         mob_segment_entity: Entity,
         mob_segment_faction: Faction,
-        player_damage: f32,
-        mob_segment_damage: f32,
+        player_damage: usize,
+        mob_segment_damage: usize,
     },
     PlayerToProjectileContact {
         player_entity: Entity,
         projectile_entity: Entity,
         projectile_faction: Faction,
-        player_damage: f32,
-        projectile_damage: f32,
+        player_damage: usize,
+        projectile_damage: usize,
     },
 
     // Mob to projectile
@@ -68,7 +68,7 @@ pub enum SortedCollisionEvent {
         projectile_entity: Entity,
         mob_faction: Faction,
         projectile_faction: Faction,
-        projectile_damage: f32,
+        projectile_damage: usize,
     },
     MobToProjectileContact {
         projectile_source: Entity,
@@ -76,7 +76,7 @@ pub enum SortedCollisionEvent {
         projectile_entity: Entity,
         projectile_faction: Faction,
         mob_faction: Faction,
-        projectile_damage: f32,
+        projectile_damage: usize,
     },
 
     // Mob segment to projectile
@@ -85,42 +85,42 @@ pub enum SortedCollisionEvent {
         projectile_entity: Entity,
         mob_segment_faction: Faction,
         projectile_faction: Faction,
-        projectile_damage: f32,
+        projectile_damage: usize,
     },
     MobSegmentToProjectileContact {
         mob_segment_entity: Entity,
         projectile_entity: Entity,
         mob_segment_faction: Faction,
         projectile_faction: Faction,
-        projectile_damage: f32,
+        projectile_damage: usize,
     },
 
     // Mob to mob
     MobToMobContact {
         mob_entity_1: Entity,
         mob_faction_1: Faction,
-        mob_damage_1: f32,
+        mob_damage_1: usize,
         mob_entity_2: Entity,
         mob_faction_2: Faction,
-        mob_damage_2: f32,
+        mob_damage_2: usize,
     },
     MobToMobSegmentContact {
         mob_entity: Entity,
         mob_faction: Faction,
-        mob_damage: f32,
+        mob_damage: usize,
         mob_segment_entity: Entity,
         mob_segment_faction: Faction,
-        mob_segment_damage: f32,
+        mob_segment_damage: usize,
     },
 
     // Mob segment to mob segment
     MobSegmentToMobSegmentContact {
         mob_segment_entity_1: Entity,
         mob_segment_faction_1: Faction,
-        mob_segment_damage_1: f32,
+        mob_segment_damage_1: usize,
         mob_segment_entity_2: Entity,
         mob_segment_faction_2: Faction,
-        mob_segment_damage_2: f32,
+        mob_segment_damage_2: usize,
     },
 
     // Projectile to projectile
